+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Start of documentation: 2023-02-16 / v1.1.0
 ### Bugfixes
 
 - fix Bug where --nocolor would not apply to --FILES output or the update information.
+- fix Bug where --FILES would have unwanted behaviour when no files have been found.
 
 
 ## [1.1.1] - 2023-02-19